api_backend/pyproject.toml
[tool.poetry]
name = "on-api"
version = "0.1.0"
description = ""
authors = ["jeanmarc"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.16.0"
fastapi = "^0.75.1"
pydantic = {extras = ["email"], version = "^1.9.0"}
SQLAlchemy = "^1.4.31"
psycopg2-binary = "^2.9.3"
SQLAlchemy-Utils = "^0.38.2"
alembic = "^1.7.6"
PyYAML = "^6.0"
requests = "^2.27.1"
databases = "^0.5.5"
bcrypt = "^3.2.0"
passlib = "^1.7.4"
types-PyYAML = "^6.0.5"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
jose = "^1.0.0"
web3 = "^5.29.0"
types-requests = "^2.27.25"
qrcode = "^7.3.1"
Pillow = "^9.2.0"
fastapi-mail = "^1.1.4"
Jinja2 = "^3.1.2"
aioredis = "^2.0.1"
stubs = "^1.0.0"
pandas = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
flake8 = "^4.0.1"
black = {version = "^21.11b1", allow-prereleases = true}
mypy = "^0.930"
pytest-cov = "^2.12"
pytest-env = "^0.6.2"
pytest-xdist = "^2.4.0"
httpx = "^0.19.0"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.18.3"


[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=0.5
  --numprocesses=auto
  --asyncio-mode=strict
  
'''

[tool.coverage.run]
omit = [
    "app/db/models/*",
    "app/api/utils/role_has_rights.py",
    "app/api/utils/user_has_roles.py",
    "app/api/v1/route_role_has_rights.py",
    "app/api/v1/route_user_has_roles.py",
    ]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"